Report the resource diagnostics from an analyzer, not the generator
A source generator runs on every keystroke in the IDE, so it has to stay as
cheap as possible. Parsing every `.resw` file of every language of a project and
cross checking them against the default language is far too expensive to do
there, and it is not what a generator is for.

The rules are therefore moved to a proper `DiagnosticAnalyzer`, which the
compiler schedules independently of the generator pipeline, runs out of process,
and which consumers can disable per rule. The check logic itself is unchanged,
it simply moved from `ReswResourceAnalyzer` to `ReswResourceRules` so that the
analyzer is a thin entry point over it, and it is now also covered by a test
that goes through `CompilationWithAnalyzers`, which is the only thing that can
catch a registration mistake.

While reviewing the pipeline: the generator held on to the `Compilation` itself,
which is a different object after every edit, so every downstream step was
invalidated on every keystroke. It now projects it to an equatable
`CompilationInfo` holding only the three things the generation actually reads,
so the pipeline is only invalidated when one of them changes.

Note that a compilation end action is skipped once the compilation already has
errors, so RESWP0009 does not appear on the build that a resource conflict makes
fail. Reporting it from the generator instead would have made it visible there,
but at a cost paid by every keystroke of every consumer, which is not a good
trade.
