+++ src/ReswPlus.SourceGenerator/Diagnostics.cs
@@ -110,12 +110,12 @@ internal static class Diagnostics
         isEnabledByDefault: true);
 
     /// <summary>
-    /// RESWP0009: two resources of the same file are generated as the same member.
+    /// RESWP0009: two resources of the same file conflict with each other.
     /// </summary>
     public static readonly DiagnosticDescriptor DuplicateResource = new(
         "RESWP0009",
-        "Duplicate resource",
-        "The resource '{0}' generates the member '{1}', which is already generated by the resource '{2}'",
+        "Conflicting resources",
+        "The resource '{0}' conflicts with the resource '{1}' declared earlier in the same file",
         ResourcesCategory,
         DiagnosticSeverity.Warning,
         isEnabledByDefault: true);
